refactor(components): migrate NotificationBanner to TypeScript
feature/typescript-messages
